refactor(schema): move schema Rule to schema package
