+++ packages/@sanity/schema/src/_exports/index.ts
@@ -2,4 +2,5 @@ import {DeprecatedDefaultSchema, Schema as NamedSchema} from '../legacy/Schema'
 
 export default DeprecatedDefaultSchema
 export const Schema = NamedSchema
+export {Rule} from '../legacy/Rule'
 export {type SchemaValidationResult} from '../sanity/typedefs'
